Move token update to Manage/ExternalLogins
