+++ Pandacap/Areas/Identity/Pages/Account/Manage/ExternalLogins.cshtml.cs
@@ -2,32 +2,32 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 #nullable disable
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using Microsoft.EntityFrameworkCore;
+using Pandacap.Data;
 
-namespace WebApplication2.Areas.Identity.Pages.Account.Manage
+namespace Pandacap.Areas.Identity.Pages.Account.Manage
 {
     public class ExternalLoginsModel : PageModel
     {
         private readonly UserManager<IdentityUser> _userManager;
         private readonly SignInManager<IdentityUser> _signInManager;
         private readonly IUserStore<IdentityUser> _userStore;
+        private readonly PandacapDbContext _context;
 
         public ExternalLoginsModel(
             UserManager<IdentityUser> userManager,
             SignInManager<IdentityUser> signInManager,
-            IUserStore<IdentityUser> userStore)
+            IUserStore<IdentityUser> userStore,
+            PandacapDbContext context)
         {
             _userManager = userManager;
             _signInManager = signInManager;
             _userStore = userStore;
+            _context = context;
         }
 
         /// <summary>
@@ -131,11 +131,63 @@ namespace WebApplication2.Areas.Identity.Pages.Account.Manage
                 return RedirectToPage();
             }
 
+            await UpdateTokensAsync(user, info);
+
             // Clear the existing external cookie to ensure a clean login process
             await HttpContext.SignOutAsync(IdentityConstants.ExternalScheme);
 
             StatusMessage = "The external login was added.";
             return RedirectToPage();
+        }
+
+        private async Task UpdateTokensAsync(IdentityUser user, ExternalLoginInfo info)
+        {
+            if (info.LoginProvider == "DeviantArt")
+            {
+                var credentials = await _context.DeviantArtCredentials
+                    .Where(c => c.Username == info.Principal.Identity.Name)
+                    .FirstOrDefaultAsync();
+                if (credentials == null)
+                {
+                    credentials = new DeviantArtCredentials
+                    {
+                        Username = info.Principal.Identity.Name
+                    };
+                    _context.DeviantArtCredentials.Add(credentials);
+                }
+                credentials.AccessToken = info.AuthenticationTokens
+                    .Where(t => t.Name == "access_token")
+                    .Select(t => t.Value)
+                    .Single();
+                credentials.RefreshToken = info.AuthenticationTokens
+                    .Where(t => t.Name == "refresh_token")
+                    .Select(t => t.Value)
+                    .Single();
+                await _context.SaveChangesAsync();
+            }
+            else if (info.LoginProvider == "Reddit")
+            {
+                var credentials = await _context.RedditCredentials
+                    .Where(c => c.Username == info.Principal.Identity.Name)
+                    .FirstOrDefaultAsync();
+                if (credentials == null)
+                {
+                    credentials = new RedditCredentials
+                    {
+                        Username = info.Principal.Identity.Name
+                    };
+                    _context.RedditCredentials.Add(credentials);
+                }
+                credentials.AccessToken = info.AuthenticationTokens
+                    .Where(t => t.Name == "access_token")
+                    .Select(t => t.Value)
+                    .Single();
+                credentials.RefreshToken = info.AuthenticationTokens
+                    .Where(t => t.Name == "refresh_token")
+                    .Select(t => t.Value)
+                    .Single();
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
